pyproject.toml
[tool.poetry]
name = "resource-monitor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psutil = "^6.0.0"
pandas = "^2.2.3"
slack-sdk = "^3.33.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
